[ty] Improve `sys.version_info` special casing
